chnaged uint 16 to unsigned for linesensorvalues

+++ include/LineSensor.h
@@ -6,12 +6,12 @@
 class LineSensor {
 private:
     Zumo32U4LineSensors lineSensors;
-    uint16_t lineSensorValues[5];
+    unsigned lineSensorValues[5];
 
 public:
     LineSensor();
     void detectLine();
-    const uint16_t* getSensorValues() const;
+    const unsigned* getSensorValues() const;
 };
 
 #endif // LINESENSOR_H
